rename base_register_decorator to be private.  Start test folder breakup.
