+++ topik/feature_transformations/_registry.py
@@ -1,6 +1,6 @@
 from functools import partial
 
-from topik.singleton_registry import BaseRegistry, base_register_decorator
+from topik.singleton_registry import BaseRegistry, _base_register_decorator
 
 
 # This subclass serves to establish a new singleon instance of functions
@@ -14,9 +14,9 @@ registered_transformers = TransformerRegistry()
 
 
 # fill in the registration function
-register = partial(base_register_decorator, TransformerRegistry)
+register = partial(_base_register_decorator, TransformerRegistry)
 
 
 # this function is the primary API for people using any registered functions.
-def run_model(model_name, input_data, **kwargs):
+def transform(tranformation_name, input_data, **kwargs):
     return TransformerRegistry()[model_name](input_data, **kwargs)
